Batched NMS
Requires a new version of NMS that returns indices sorted by score. This implementation is 15% faster than the previous one during inference, and a bit faster during training. Should keep an eye on about higher memory usage, even though for the standard models this was not the case, and it also used slightly less memory